+++ client/src/components/people/people.module.css
@@ -1,0 +1,5 @@
+.plusContainer{
+	position:fixed;
+	bottom:20px;
+    right:20px;
+}
