+++ client/src/components/profile/profile.module.css
@@ -1,9 +1,3 @@
-.plusContainer{
-	position:fixed;
-	bottom:20px;
-    right:20px;
-}
-
 .imageContainer {
 	width: 200px;
 	height: 200px;
